requirements.txt
PyMuPDF
pytesseract
opencv-python
numpy
pillow
dominate
tqdm
pytest
beautifulsoup4
